refactor(core): adopt ImageDataLike in pipeline utils (phase 8)
Move finalize.ts, final-composite.ts, auto-crop.ts, finalize-result.ts into nukebg-core/src/pipeline and replace all 7 'new ImageData(...)' call sites with createImageDataLike — core no longer depends on the ImageData DOM global. Core tests construct plain objects (deleted 5 ImageData polyfills). App boundary gains toImageData() helpers (ar-app, batch-orchestrator) to convert plain ImageDataLike back to native ImageData for canvas putImageData. X.7: finalize-chain test moved to core; pending-timers stays in app.

Closes #273
