+++ packages/nukebg-app/src/components/ar-editor-advanced.ts
@@ -29,7 +29,7 @@
  */
 
 import { createRmbg14Loader, type Rmbg14Loader } from '../refine/loaders/rmbg14';
-import { refineEdges } from '../pipeline/finalize';
+import { refineEdges } from 'nukebg-core/pipeline/finalize';
 import { SamRefiner } from '../controllers/sam-refiner';
 import { processRoi } from '../refine/roi-process';
 import { rasterizePolygon } from '../refine/roi-process';
