+++ packages/nukebg-core/src/pipeline/auto-crop.ts
@@ -1,27 +1,15 @@
+import { createImageDataLike } from '../types/image-data-like.js';
+import type { ImageDataLike } from '../types/image-data-like.js';
+
 /**
  * Tight bounding-box crop on the subject (any pixel with alpha above
  * `alphaThreshold`).
  *
- * The pipeline preserves the original canvas size — every transparent
- * pixel outside the subject is still in the output buffer, just with
- * α=0. That's good for the before/after slider (canvases align) but
- * wasteful for export: a 1920×1080 photo with a 200×200 subject ships a
- * 1920×1080 PNG.
- *
- * `autoCropToSubject` returns a fresh ImageData clipped to the subject's
- * bbox so the downloaded file is just the meaningful pixels — sized
- * right for emotes, stickers, profile pictures.
- *
- * Edge cases:
- *   - all-transparent: returns the input unchanged. There's no subject
- *     to crop to and a 0×0 PNG is useless.
- *   - all-opaque (bbox == full image): returns the input unchanged. No
- *     allocation cost paid for a no-op.
- *   - threshold default 1: any non-zero α counts as subject. The
- *     pipeline's quintic sharpenAlpha already kills halos to α=0, so
- *     α>0 lands on actual antialiased edge — cropping there preserves
- *     the soft transition on a clean transparent background.
+ * Moved from packages/nukebg-app/src/utils/auto-crop.ts in Phase 8.
+ * `new ImageData(out, cw, ch)` replaced with `createImageDataLike(out, cw, ch)`.
+ * No DOM globals used.
  */
+
 export interface AutoCropOptions {
   /** Pixels with α >= this count as subject. Default 1 (any non-zero α). */
   alphaThreshold?: number;
@@ -29,7 +17,7 @@ export interface AutoCropOptions {
   padding?: number;
 }
 
-export function autoCropToSubject(img: ImageData, options: AutoCropOptions = {}): ImageData {
+export function autoCropToSubject(img: ImageDataLike, options: AutoCropOptions = {}): ImageDataLike {
   const threshold = options.alphaThreshold ?? 1;
   const padding = options.padding ?? 0;
   const { width: w, height: h, data } = img;
@@ -42,7 +30,7 @@ export function autoCropToSubject(img: ImageData, options: AutoCropOptions = {})
   for (let y = 0; y < h; y++) {
     const row = y * w;
     for (let x = 0; x < w; x++) {
-      if (data[(row + x) * 4 + 3] >= threshold) {
+      if ((data[(row + x) * 4 + 3] ?? 0) >= threshold) {
         if (x < minX) minX = x;
         if (x > maxX) maxX = x;
         if (y < minY) minY = y;
@@ -51,7 +39,7 @@ export function autoCropToSubject(img: ImageData, options: AutoCropOptions = {})
     }
   }
 
-  // No subject pixels at all — return unchanged. A 0×0 PNG is useless.
+  // No subject pixels at all — return unchanged.
   if (maxX < 0) return img;
 
   const x0 = Math.max(0, minX - padding);
@@ -70,5 +58,5 @@ export function autoCropToSubject(img: ImageData, options: AutoCropOptions = {})
     const dstRow = y * cw * 4;
     out.set(data.subarray(srcRow, srcRow + cw * 4), dstRow);
   }
-  return new ImageData(out, cw, ch);
+  return createImageDataLike(out, cw, ch);
 }
