+++ packages/nukebg-core/tests/pipeline/finalize-chain.test.ts
@@ -4,42 +4,21 @@ import {
   fillSubjectHoles,
   promoteSpeckleAlpha,
 } from '../../src/pipeline/finalize';
+import { createImageDataLike } from '../../src/types/image-data-like';
+import type { ImageDataLike } from '../../src/types/image-data-like';
 
-// ImageData polyfill for happy-dom.
-if (typeof globalThis.ImageData === 'undefined') {
-  (globalThis as unknown as { ImageData: unknown }).ImageData = class ImageData {
-    data: Uint8ClampedArray;
-    width: number;
-    height: number;
-    constructor(
-      dataOrWidth: Uint8ClampedArray | number,
-      widthOrHeight: number,
-      maybeHeight?: number,
-    ) {
-      if (dataOrWidth instanceof Uint8ClampedArray) {
-        this.data = dataOrWidth;
-        this.width = widthOrHeight;
-        this.height = maybeHeight ?? dataOrWidth.length / (widthOrHeight * 4);
-      } else {
-        this.width = dataOrWidth;
-        this.height = widthOrHeight;
-        this.data = new Uint8ClampedArray(this.width * this.height * 4);
-      }
-    }
-  };
-}
+// Core runs in Node — NO ImageData global, NO polyfill.
+// All inputs are plain ImageDataLike objects.
 
 // End-to-end topology cleanup chain: the same composition ar-app.ts applies
 // on the PHOTO path. Guards that chaining the three passes produces a
 // coherent output — no enclosed α=0 holes, no interior specks, no
 // disconnected α>0 blobs.
-const runChain = (img: ImageData): ImageData =>
+const runChain = (img: ImageDataLike): ImageDataLike =>
   promoteSpeckleAlpha(fillSubjectHoles(dropOrphanBlobs(img)));
 
-// Build a synthetic image containing every defect class the chain must cure.
-const makeDefectiveSubject = () => {
-  const w = 40,
-    h = 40;
+const makeDefectiveSubject = (): ImageDataLike => {
+  const w = 40, h = 40;
   const data = new Uint8ClampedArray(w * h * 4);
   // Solid body at (5,5)-(34,34): α=255, RGB=(200,100,50).
   for (let y = 5; y <= 34; y++) {
@@ -57,19 +36,17 @@ const makeDefectiveSubject = () => {
   data[(20 * w + 20) * 4 + 3] = 0;
   // Defect 3: α-intermediate speck at (15,15) surrounded by α=255 neighbors.
   data[(15 * w + 15) * 4 + 3] = 150;
-  return new ImageData(data, w, h);
+  return createImageDataLike(data, w, h);
 };
 
-// Count enclosed α=0 clusters — topology oracle used as the output invariant.
-const countEnclosedHoles = (img: ImageData): number => {
+const countEnclosedHoles = (img: ImageDataLike): number => {
   const { data, width: w, height: h } = img;
   const n = w * h;
   const queue = new Int32Array(n);
   const bg = new Uint8Array(n);
-  let head = 0,
-    tail = 0;
+  let head = 0, tail = 0;
   const seed = (i: number) => {
-    if (data[i * 4 + 3] === 0 && !bg[i]) {
+    if ((data[i * 4 + 3] ?? 0) === 0 && !bg[i]) {
       bg[i] = 1;
       queue[tail++] = i;
     }
@@ -83,11 +60,10 @@ const countEnclosedHoles = (img: ImageData): number => {
     seed(y * w + w - 1);
   }
   while (head < tail) {
-    const i = queue[head++];
-    const x = i % w,
-      y = (i - x) / w;
+    const i = queue[head++] ?? 0;
+    const x = i % w, y = (i - x) / w;
     const push = (ni: number) => {
-      if (data[ni * 4 + 3] === 0 && !bg[ni]) {
+      if ((data[ni * 4 + 3] ?? 0) === 0 && !bg[ni]) {
         bg[ni] = 1;
         queue[tail++] = ni;
       }
@@ -100,18 +76,16 @@ const countEnclosedHoles = (img: ImageData): number => {
   const seen = new Uint8Array(n);
   let holes = 0;
   for (let i = 0; i < n; i++) {
-    if (data[i * 4 + 3] !== 0 || bg[i] || seen[i]) continue;
+    if ((data[i * 4 + 3] ?? 0) !== 0 || bg[i] || seen[i]) continue;
     holes++;
-    head = 0;
-    tail = 0;
+    head = 0; tail = 0;
     queue[tail++] = i;
     seen[i] = 1;
     while (head < tail) {
-      const j = queue[head++];
-      const x = j % w,
-        y = (j - x) / w;
+      const j = queue[head++] ?? 0;
+      const x = j % w, y = (j - x) / w;
       const push = (nj: number) => {
-        if (data[nj * 4 + 3] === 0 && !bg[nj] && !seen[nj]) {
+        if ((data[nj * 4 + 3] ?? 0) === 0 && !bg[nj] && !seen[nj]) {
           seen[nj] = 1;
           queue[tail++] = nj;
         }
@@ -125,19 +99,18 @@ const countEnclosedHoles = (img: ImageData): number => {
   return holes;
 };
 
-// Count interior α-intermediate specks with ≥80% opaque neighbors in 5×5.
-const countInteriorSpecks = (img: ImageData): number => {
+const countInteriorSpecks = (img: ImageDataLike): number => {
   const { data, width: w, height: h } = img;
   let count = 0;
   for (let y = 2; y < h - 2; y++) {
     for (let x = 2; x < w - 2; x++) {
-      const a = data[(y * w + x) * 4 + 3];
+      const a = data[(y * w + x) * 4 + 3] ?? 0;
       if (a === 0 || a === 255) continue;
       let opaque = 0;
       for (let dy = -2; dy <= 2; dy++) {
         for (let dx = -2; dx <= 2; dx++) {
           if (dx === 0 && dy === 0) continue;
-          if (data[((y + dy) * w + (x + dx)) * 4 + 3] >= 240) opaque++;
+          if ((data[((y + dy) * w + (x + dx)) * 4 + 3] ?? 0) >= 240) opaque++;
         }
       }
       if (opaque / 24 >= 0.8) count++;
@@ -146,22 +119,19 @@ const countInteriorSpecks = (img: ImageData): number => {
   return count;
 };
 
-describe('finalize chain (dropOrphanBlobs → fillSubjectHoles → promoteSpeckleAlpha)', () => {
+describe('finalize chain (dropOrphanBlobs → fillSubjectHoles → promoteSpeckleAlpha) [core]', () => {
   it('kills disconnected orphan α>0 blobs', () => {
     const out = runChain(makeDefectiveSubject());
-    // Orphan at (38,38) must be α=0.
     expect(out.data[(38 * 40 + 38) * 4 + 3]).toBe(0);
   });
 
   it('fills small topologically enclosed α=0 holes', () => {
     const out = runChain(makeDefectiveSubject());
-    // Enclosed hole at (20,20) must now be opaque.
     expect(out.data[(20 * 40 + 20) * 4 + 3]).toBe(255);
   });
 
   it('promotes α-intermediate specks surrounded by opaque body', () => {
     const out = runChain(makeDefectiveSubject());
-    // Speck at (15,15) must now be α=255.
     expect(out.data[(15 * 40 + 15) * 4 + 3]).toBe(255);
   });
 
@@ -180,8 +150,8 @@ describe('finalize chain (dropOrphanBlobs → fillSubjectHoles → promoteSpeckl
     const w = 40;
     for (let y = 5; y <= 34; y++) {
       for (let x = 5; x <= 34; x++) {
-        if (x === 20 && y === 20) continue; // the enclosed hole, verified above
-        if (x === 15 && y === 15) continue; // the promoted speck, verified above
+        if (x === 20 && y === 20) continue;
+        if (x === 15 && y === 15) continue;
         expect(out.data[(y * w + x) * 4 + 3]).toBe(255);
       }
     }
@@ -191,7 +161,7 @@ describe('finalize chain (dropOrphanBlobs → fillSubjectHoles → promoteSpeckl
     const input = makeDefectiveSubject();
     const rgbSnapshot: number[] = [];
     for (let i = 0; i < input.data.length; i += 4) {
-      rgbSnapshot.push(input.data[i], input.data[i + 1], input.data[i + 2]);
+      rgbSnapshot.push(input.data[i] ?? 0, input.data[i + 1] ?? 0, input.data[i + 2] ?? 0);
     }
     const out = runChain(input);
     for (let i = 0, j = 0; i < out.data.length; i += 4, j += 3) {
@@ -201,7 +171,7 @@ describe('finalize chain (dropOrphanBlobs → fillSubjectHoles → promoteSpeckl
     }
   });
 
-  it('does not mutate the input ImageData', () => {
+  it('does not mutate the input ImageDataLike', () => {
     const input = makeDefectiveSubject();
     const snapshot = Array.from(input.data);
     runChain(input);
